#0: split tg test into another file
